feat: enable password changes in coder-sdk
* Allow modification of passwords and the temporary_password flag
  for users managed by the built-in authentication provider.
* Split unit tests into separate files corresponding to the
  implementations.
